Migrate to allure-ruby-api.

+++ lib/allure-rspec/formatter.rb
@@ -1,6 +1,5 @@
-require 'pathname'
-require 'uuid'
 require 'rspec/core/formatters/base_formatter'
+require 'fileutils'
 
 module AllureRSpec
 
@@ -9,7 +8,7 @@ module AllureRSpec
     ALLOWED_LABELS = [:feature, :story, :severity, :language, :framework]
 
     def example_failed(example)
-      AllureRSpec::Builder.stop_test(
+      AllureRubyApi::Builder.stop_test(
           example.metadata[:example_group][:description_args].first,
           example.metadata[:description],
           example.metadata[:execution_result].merge(
@@ -21,17 +20,17 @@ module AllureRSpec
     end
 
     def example_group_finished(group)
-      AllureRSpec::Builder.stop_suite(group.metadata[:example_group][:description_args].first)
+      AllureRubyApi::Builder.stop_suite(group.metadata[:example_group][:description_args].first)
       super
     end
 
     def example_group_started(group)
-      AllureRSpec::Builder.start_suite(group.metadata[:example_group][:description_args].first, labels(group))
+      AllureRubyApi::Builder.start_suite(group.metadata[:example_group][:description_args].first, labels(group))
       super
     end
 
     def example_passed(example)
-      AllureRSpec::Builder.stop_test(
+      AllureRubyApi::Builder.stop_test(
           example.metadata[:example_group][:description_args].first,
           example.metadata[:description],
           example.metadata[:execution_result].merge(:caller => example.metadata[:caller])
@@ -46,31 +45,21 @@ module AllureRSpec
     def example_started(example)
       suite = example.metadata[:example_group][:description_args].first
       test = example.metadata[:description]
-      AllureRSpec::Builder.start_test(suite, test, labels(example))
+      AllureRubyApi::Builder.start_test(suite, test, labels(example))
       super
     end
 
     def start(example_count)
       dir = Pathname.new(AllureRSpec::Config.output_dir)
-      FileUtils.rm_rf(dir)
+      if AllureRSpec::Config.clean_dir?
+        FileUtils.rm_rf(dir)
+      end
+      FileUtils.mkdir_p(dir)
+      super
     end
 
     def stop
-      AllureRSpec::Builder.each_suite_build do |suite, xml|
-        dir = Pathname.new(AllureRSpec::Config.output_dir)
-        FileUtils.mkdir_p(dir)
-        out_file = dir.join("#{UUID.new.generate}-testsuite.xml")
-        File.open(out_file, 'w+') do |file|
-          file.write(xml)
-        end
-
-        xsd = Nokogiri::XML::Schema(File.read(Pathname.new(File.dirname(__FILE__)).join("../../allure-model.xsd")))
-        doc = Nokogiri::XML(File.read(out_file))
-
-        xsd.validate(doc).each do |error|
-          $stderr.puts error.message
-        end
-      end
+      AllureRubyApi::Builder.build!
       super
     end
